pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.10"
flake8 = "^7.0.0"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
PyYAML = "^6.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
